requirements.txt
biopython==1.81
numpy==1.26.2
scikit-learn==1.5.0
tqdm==4.66.3
biotite~=1.0.1
click~=8.1.8